requirements.txt
fastapi==0.70.0
uvicorn[standard]==0.15.0
gunicorn==20.1.0
python-dotenv==0.19.1
sqlmodel==0.0.4
psycopg2-binary==2.9.1
click==8.0.3
typer==0.4.0
tqdm==4.62.3
spacy==3.1.4
https://github.com/explosion/spacy-models/releases/download/ru_core_news_md-3.1.0/ru_core_news_md-3.1.0-py3-none-any.whl
rank_bm25==0.2.1
pystemmer==2.0.1
gensim==3.8.3
compress-fasttext==0.0.7
faiss-cpu==1.7.1.post2
numpy~=1.21.3
sqlalchemy~=1.4.26
scikit-learn==1.0.1
